+++ zia/outputs.tf
@@ -1,4 +1,0 @@
-output "zia_dlp_web_rules_resource_zia_dlp_web_rules_1112360_id" {
-  value = "${zia_dlp_web_rules.resource_zia_dlp_web_rules_1112360.id}"
-}
-